requirements.txt
fastapi
pydantic
python-dotenv
uvicorn
sentence-transformers
bitsandbytes
langchain
faiss_cpu
torch
accelerate
